+++ src/resolvers/metadataStandard.ts
@@ -8,8 +8,7 @@ import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError
 import { ResearchDomain } from '../models/ResearchDomain';
 import { GraphQLError } from 'graphql';
 import { PaginationOptionsForCursors, PaginationOptionsForOffsets, PaginationType } from '../types/general';
-import { isNullOrUndefined } from '../utils/helpers';
-import {formatISO9075} from "date-fns";
+import { isNullOrUndefined, normaliseDate } from '../utils/helpers';
 
 export const resolvers: Resolvers = {
   Query: {
@@ -256,10 +255,10 @@ export const resolvers: Resolvers = {
       return [];
     },
     created: (parent: MetadataStandard) => {
-      return formatISO9075(new Date(parent.created));
+      return normaliseDate(parent.created);
     },
     modified: (parent: MetadataStandard) => {
-      return formatISO9075(new Date(parent.modified));
+      return normaliseDate(parent.modified);
     }
   },
 };