+++ src/resolvers/repository.ts
@@ -6,10 +6,13 @@ import { MyContext } from '../context';
 import { isAdmin, isAuthorized, isSuperAdmin } from '../services/authService';
 import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError } from '../utils/graphQLErrors';
 import { ResearchDomain } from '../models/ResearchDomain';
-import { isNullOrUndefined, stringToEnumValue } from '../utils/helpers';
+import {
+  isNullOrUndefined,
+  normaliseDate,
+  stringToEnumValue
+} from '../utils/helpers';
 import { GraphQLError } from 'graphql';
 import { PaginationOptionsForCursors, PaginationOptionsForOffsets, PaginationType } from '../types/general';
-import {formatISO9075} from "date-fns";
 
 export const resolvers: Resolvers = {
   Query: {
@@ -287,10 +290,10 @@ export const resolvers: Resolvers = {
       );
     },
     created: (parent: Repository) => {
-      return formatISO9075(new Date(parent.created));
+      return normaliseDate(parent.created);
     },
     modified: (parent: Repository) => {
-      return formatISO9075(new Date(parent.modified));
+      return normaliseDate(parent.modified);
     }
   },
 };